core: wifi: Make wifi service dinamically get first wireless interface available
This commit makes the wifi service dinamically get the first available
interface under `/var/wpa_supplicant`. With this, the service runs without
changes in most machines.
For machines with more than one interface, there's now the option to connect
to the desired interface with the --socket argument. This is the case for
companion, as Raspbian by default has also a WiFi-Direct interface, so we
explicitly use `wlan0`.