core: ardupilot-manager: Introduce `FlightController` type
`FlightController` is used to abstract board identification.
With it, `FlightControllerType` is deprecated in favor of `Platform`, which has the same information.

Since we only support Pixhawk1 by the time, all serial boards are treated like so, but this will need
to be changed in the future, differentiating between generic serial boards and known ones.

NavigatorR3, which was a FlightControllerType previously, is added to the available platforms,
with the same string representation as default Navigator.

Signed-off-by: Rafael Araujo Lehmkuhl <rafael.lehmkuhl93@gmail.com>