core: ardupilot-manager: Refactor master endpoint logic
Master endpoint was being stored in two places, Manager and AbstractRouter, having no single source of truth.
An example where the problem is clear is if one set's the master endpoint (on Manager) and restart it. The used master endpoint will be the old one (stored on AbstractRouter).

Making the master endpoint stored only on AbstractRouter, this problem cease to exist.