core: ardupilot-manager: Deprecate `Platform.Undefined` type
When a platform is not defined by using the `Undefined` enum option, it mandates that everyone using a variable defined as a Platform to check if its indeed a valid Platform option, or just an Undefined.

By removing this enum option and explicitly using None when the platform is not defined we make things clear. If no platform is defined (None) the static type-checker will point this upfront or a error will be raised, informing the one using the variable value.