bootstrap, core/start-blueos-core: Sync host and container resolv.conf
This is a workaround to make the container's `resolv.conf` synchronized
with the host's `resolv.conf`. To do that, we disallow docker from
managing the container's DNS nameservers by unmounting the container's
`/etc/resolv.conf`, and soft linking it to the binded host's
`/etc/resolv.conf`.  So for this to work, the container's
`/etc/resolv.conf.host` file should be binded (read-only) to the host's
`/etc/resolv.conf`, for our system, it's configured in
`bootstrap/startup.json.default`, and a patch is added in
`core/tools/blueos_startup_update` so systems with older bootstrap
versions can also work with it.

One detail is that because docker uses `mount` to bind the host's file
to the container's file, if the file-descriptor of the file changes, the
bind will be lost. To prevent this, the file being binded must be a hard
link to the actual `/etc/resolv.conf`, and that hard link file is the
one being binded. This hard link is being done in
`core/tools/blueos_startup_update` as `/etc/resolv.conf.host`.

To briefly understand why it is
necessary, when the docker starts a container, it generates the
container `/etc/resolv.conf` based on the host's one, or some docker
configuration, however, during the container's lifetime, if the host's
DNS configuration changes, the container's DNS doesn't follow it, making
the container (possibly) unable to resolve DNSs correctly, failing the
BlueOS Helper's internet connection check. By syncing the container's
resolv.conf with the host's one, we make sure the container's DNS
configuration follows the host's changes, thus, ensuring the container
is able to resolve DNSs correctly.