Fix nginx.conf syntax that I broke
Undo my introduction of invalid syntax in #2913, which breaks blueos.

It turns out that you can't start a URL with a leading colon. The only reason I thought it gave the same results is because `service nginx reload` bailed when loading the new config, but left the server running with the same state.